change to autotools build system
With  respect  to  usage  on  UNIX  based  operating  systems,  it  is
advantageous to  switch to the  GNU build system autotools  since this
offers superior  abstraction from  OS specific library  handling. This
requires to  alter the original  Makefiles provided by  the tinyscheme
project. In this respect the directory structure is changed as well to
use the very  same flat structure in which the  tinyscheme project was
originally provided.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>